+++ backend/helpus_donation/helpus/src/main/java/com/ssafy/helpus/donation/repository/DonationImageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DonationImageRepository extends JpaRepository<DonationImage, Integer> {
+public interface DonationImageRepository extends JpaRepository<DonationImage, Long> {
 }
